server/requirements.txt
# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
# langchain-google-genai

google-generativeai

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.0.0

# FastAPI backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
poppler-utils

# OCR for images and scanned PDFs
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# Web scraping for lawyer search
httpx>=0.27.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async support
aiofiles>=24.1.0

# Caching for performance
cachetools>=5.5.0
